Move test/proptest/__init__.py to setup.py
