+++ tests/proptest/realtime/test_prop_BusGroup.py
@@ -10,7 +10,7 @@ import supriya.realtime
 import supriya.synthdefs
 from supriya import CalculationRate
 from supriya.exceptions import BusNotAllocated, IncompatibleRate
-from tests.proptest import get_control_test_groups, hp_global_settings
+from tests.proptest.setup import get_control_test_groups, hp_global_settings
 
 
 @pytest.fixture(autouse=True)
